pyproject.toml
[project]
name = "openai-agent-sdk-standard-project"
version = "1.0.0"
description = "Enterprise-grade Agentic AI implementation following OpenAI Agent SDK best practices"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.13.4",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.0.280",
]
